Footer/Nav/Ad/LocationTab/AboutTab
Finished responsive/design for both Location tab and About tab.
Finished the footer.

Discovered new bugs: When navbar is small(cell phone size) it will not
drop down on any of the pages except for when you are on the home page.

I will also need help with doing the pop up “mail” add because I am
having trouble understand exactly what is going on.

The only thing I have left to do is make a “mail” ad for people to
click on, make all of the social network buttons into links, and figure
out what the heck is going on with the nav. I also need to make the
“ad” look better. Almost done.