Topbanner, Imagelinks, termspage
All of the links to social media are working now. The logo at the top
will take you to the main page. I made the Terms of service page and
all that is how it is supposed to work. I also got the top banner
working(mostly), it still has some responsive issues. I also made the
background for the mailing list.